rag_monitoring_service/requirements.txt
openai
psycopg2-binary
sqlalchemy
python-dotenv
scikit-learn
numpy
